Kompilacja na studentsie rzucała WARNIG.
git-svn-id: http://svn2.assembla.com/svn/mimhex@76 e2cb09e9-51e2-4e97-b92c-ac25115bd917

+++ implementation/mcts_node.cpp
@@ -167,7 +167,7 @@ void MCTSNode::Update(bool won, uint* begin, uint* end) {
 
     if (Switches::Rave() && !IsLeaf()) {
 
-        ASSERT(end - begin == count);
+        ASSERT(end - begin == (int) count);
         /// Iterate over opponent moves.
         for (uint* it = begin; it != end && it != end + 1; it += 2) {
             ASSERT(*it != loc.GetPos());
